@@ -1,214 +1,11 @@
 import textwrap
 
-
-class EditBlockPrompts:
-    @staticmethod
-    def format_main_system(fence: tuple[str, str]) -> str:
-        return textwrap.dedent(
-            """\
-            Act as an expert software developer. You are diligent and tireless!
-            You NEVER leave comments describing code without implementing it!
-            You always COMPLETELY IMPLEMENT the needed code! Always use best practices when coding.
-            Respect and use existing conventions, libraries, etc that are already present in the code base.
-
-            Take requests for changes to the supplied code. If the request is ambiguous, ask questions.
-
-            Once you understand the request you MUST:
-            1. Only suggest changes to a *read-write* files.
-            2. Think step-by-step needed changes.
-            3. Describe each change with a *SEARCH/REPLACE block* per the examples below.
-
-            All changes to files must use this *SEARCH/REPLACE block* format.
-
-            # Example conversation 1:
-
-            ## USER: These are the *read-write* files:
-
-            mathweb/flask/app.py
-            {fence[0]}python
-            from flask import Flask
-
-            app = Flask(__name__)
-
-            @app.route('/div/<int:n>/<int:d>')
-            def divide(n, d):
-                if d == 0:
-                    return 'Error: Division by zero'
-                else:
-                    result = n / d
-                    return str(result)
-
-            def factorial(n):
-                "compute factorial"
-
-                if n == 0:
-                    return 1
-                else:
-                    return n * factorial(n-1)
-
-            @app.route('/factorial/<int:n>')
-            def get_factorial(n):
-                return str(factorial(n))
-
-            if __name__ == '__main__':
-                print("Starting...")
-                app.run()
-            {fence[1]}
-
-            ## ASSISTANT: Ok.
-
-            ## USER: Change get_factorial() to use math.factorial
-
-            ## ASSISTANT:
-
-            Here are the *SEARCH/REPLACE* blocks:
-
-            mathweb/flask/app.py
-            {fence[0]}python
-            <<<<<<< SEARCH
-            from flask import Flask
-            =======
-            import math
-            from flask import Flask
-            >>>>>>> REPLACE
-            {fence[1]}
-
-            mathweb/flask/app.py
-            {fence[0]}python
-            <<<<<<< SEARCH
-            def factorial(n):
-                "compute factorial"
-
-                if n == 0:
-                    return 1
-                else:
-                    return n * factorial(n-1)
-
-            =======
-            >>>>>>> REPLACE
-            {fence[1]}
-
-            mathweb/flask/app.py
-            {fence[0]}python
-            <<<<<<< SEARCH
-                return str(factorial(n))
-            =======
-                return str(math.factorial(n))
-            >>>>>>> REPLACE
-            {fence[1]}
-
-            # Example conversation 2:
-
-            ## USER: These are the *read-write* files:
-
-            main.py
-            {fence[0]}python
-            def hello():
-                "print a greeting"
-
-                print("hello")
-
-            if __name__ == '__main__':
-                hello()
-            {fence[1]}
-
-            ## ASSISTANT: Ok.
-
-            ## USER: Refactor hello() into its own file.
-
-            ## ASSISTANT:
-
-            Here are the *SEARCH/REPLACE* blocks:
-
-            hello.py
-            {fence[0]}python
-            <<<<<<< SEARCH
-            =======
-            def hello():
-                "print a greeting"
-
-                print("hello")
-            >>>>>>> REPLACE
-            {fence[1]}
-
-            main.py
-            {fence[0]}python
-            <<<<<<< SEARCH
-            def hello():
-                "print a greeting"
-
-                print("hello")
-            =======
-            from hello import hello
-            >>>>>>> REPLACE
-            {fence[1]}
-
-            # Rules
-            {system_reminder}"""
-        ).format(
-            fence=fence,
-            system_reminder=EditBlockPrompts.format_system_reminder(fence=fence),
-        )
-
-    @staticmethod
-    def format_system_reminder(fence: tuple[str, str]) -> str:
-        return textwrap.dedent(
-            """\
-            Every *SEARCH/REPLACE block* must use this format:
-            1. The file path alone on a line, eg: main.py
-            2. The opening fence and code language, eg: {fence[0]}python
-            3. The start of search block: <<<<<<< SEARCH
-            4. A contiguous chunk of lines to search for in the existing source code
-            5. The dividing line: =======
-            6. The lines to replace into the source code
-            7. The end of the replace block: >>>>>>> REPLACE
-            8. The closing fence: {fence[1]}
-
-            Every *SEARCH* section must *EXACTLY MATCH* the existing source code, character for character, including all comments, docstrings, etc.
-
-            Include *ALL* the code being searched and replaced!
-
-            Only *SEARCH/REPLACE* files that are *read-write*.
-
-            To move code within a file, use 2 *SEARCH/REPLACE* blocks: 1 to delete it from its current location, 1 to insert it in the new location.
-
-            If you want to put code in a new file, use a *SEARCH/REPLACE block* with:
-            - A new file path, including dir name if needed
-            - An empty `SEARCH` section
-            - The new file's contents in the `REPLACE` section
-
-            You are diligent and tireless!
-            You NEVER leave comments describing code without implementing it!
-            You always COMPLETELY IMPLEMENT the needed code!"""
-        ).format(fence=fence)
-
-    @staticmethod
-    def format_files_to_change(files_content: str) -> str:
-        return textwrap.dedent(
-            "These are the *read-write* files:\n{files_content}"
-        ).format(files_content=files_content)
-
-    @staticmethod
-    def format_refactor_example(content: str) -> str:
-        return textwrap.dedent(
-            "You can use the following example to make the refactor:\n{content}"
-        ).format(content=content)
-
-    @staticmethod
-    def format_diff(content: str) -> str:
-        return textwrap.dedent(
-            "Use this unified diff as baseline to make the refactor:\n{content}"
-        ).format(content=content)
+from codebase.models import RepositoryFile
 
 
 class ReplacerPrompts:
     @staticmethod
-    def format_default_msg(
-        original_snippet: str,
-        replacement_snippet: str,
-        content: str,
-        commit_message: str,
-    ):
+    def format_default_msg(original_snippet: str, replacement_snippet: str, content: str, commit_message: str):
         return textwrap.dedent(
             """\
             Replace the following snippet:
@@ -244,7 +41,7 @@ class ReplacerPrompts:
         )
 
 
-class SimpleRefactorPrompts:
+class RefactorPrompts:
     @staticmethod
     def format_system():
         """
@@ -295,16 +92,79 @@ class SimpleRefactorPrompts:
         """
         Format the files to change for the user prompt.
         """
-        return textwrap.dedent(
-            "These are the *read-write* files:\n{files_content}"
-        ).format(files_content=files_content)
-
+        return textwrap.dedent("These are the code snippets to refactor:\n{files_content}").format(
+            files_content=files_content
+        )
 
     @staticmethod
-    def format_refactor_example(content: str) -> str:
+    def format_refactor_example(file_content: str) -> str:
         """
         Format the refactor example for the user prompt.
         """
+        return textwrap.dedent("You can use the following code example to make the refactor:\n{content}").format(
+            content=file_content
+        )
+
+    @staticmethod
+    def repository_files_to_str(repository_files: list[RepositoryFile]) -> str:
+        """
+        Get the content of the files in the repository to be used in a prompt.
+        """
+        return "\n".join([
+            textwrap.dedent(
+                """\
+                file-path: {file_path}
+                <code>
+                {content}
+                </code>"""
+            ).format(file_path=repo_file.file_path, content=repo_file.content)
+            for repo_file in repository_files
+        ])
+
+
+class DiffRefactorPrompts(RefactorPrompts):
+    @staticmethod
+    def format_system():
+        """
+        Format the system prompt for the task.
+        """
         return textwrap.dedent(
-            "You can use the following example to make the refactor:\n{content}"
+            """\
+            You are an exceptional senior software engineer that is responsible for code refactoring.
+            Given the available tools and below task, which corresponds to an important step to code refactoring,
+            convert the task into code.
+            It's absolutely vital that you completely and correctly execute your task.
+
+            When the task is complete, reply with "<DONE>"
+            If you are unable to complete the task, also reply with "<DONE>"
+
+            <guidelines>
+            - Write code by calling the available tools.
+            - Ensure the code is valid and executable after applying changes.
+            - Maintain correct code padding, spacing, and indentation.
+            - Do not make extraneous changes to the code or whitespace that are not related to the intent of the change.
+            </guidelines>"""
+        )
+
+    @staticmethod
+    def format_diff(content: str) -> str:
+        return textwrap.dedent(
+            """\
+            <task>
+            - Refactor the code by applying this unified diff.
+            - The diff is a patch that contains the changes to be made to the code.
+            </task>
+
+            <diff>
+            {content}
+            </diff>
+
+            You must complete the task.
+
+            - Think out loud step-by-step before you start writing code.
+            - Do not just add a comment or leave a TODO; you must write functional code.
+            - Import libraries and modules in their own step.
+            - Carefully review your code and ensure it is formatted correctly.
+
+            You must use the provided tools/functions to do so."""
         ).format(content=content)
